@@ -1,4 +1,4 @@
-ruby '2.3.0'
-source "https://rubygems.org"
+ruby '2.6.3'
+source 'https://rubygems.org'
 
 gemspec
