@@ -1,6 +1,6 @@
-ENV["RAILS_ENV"] = "test"
+ENV['RAILS_ENV'] = 'test'
 
-#require 'rspec/rails'
+# require 'rspec/rails'
 require 'i18n-spec'
 require 'i18n/core_ext/hash'
 require 'active_support'
@@ -18,6 +18,6 @@ module RbConfig
     config.active_support.deprecation = :stderr
   end
 end
- 
+
 # Initialize the application
 RbConfig::Application.initialize!
